+++ app/utils/bpmn-download-url.js
@@ -1,5 +1,3 @@
-export default function generateBpmnDownloadUrl(fileId, fileName) {
-  let url = `/files/${fileId}/download`;
-  if (fileName) url += `?name=${fileName}`;
-  return url;
+export default function generateBpmnDownloadUrl(fileId) {
+  return `/files/${fileId}/download`;
 }
